nnue_train: score_clamp_abs を i16 型化し縮小 cast を型で排除、clamp>=drop を reject
full-diff 再レビューの指摘対応:
- CLI を Option<i16> + clap range(1..) にし、loader までの全経路で縮小 cast
  (wrap) を型レベルで不可能に。eval-only の手動範囲チェックは不要になり削除
- validate() に clamp >= drop の整合性チェックを追加 (drop 生存者は全て
  |score| < drop なので clamp が一度も発火しない silent no-op 構成を弾く)
- 手書き if/else 飽和を Ord::clamp 一行に置換 (dataloader / validation)
- 既存 drop テスト 3 件の冗長な field 初期化を除去

+++ bins/nnue_train/src/cli.rs
@@ -190,11 +190,12 @@ pub(crate) struct Cli {
     pub(crate) score_drop_abs: Option<i32>,
 
     /// Saturate the teacher score of surviving positions to `[-N, N]` (applied
-    /// after the `--score-drop-abs` filter, so e.g. mate stamps are still
-    /// dropped, not clamped). Useful to normalise teacher files whose encode
-    /// variants clip at different ceilings. Must be in `[1, 32767]`.
-    #[arg(long, global = true)]
-    pub(crate) score_clamp_abs: Option<i32>,
+    /// after the `--score-drop-abs` filter when set, so e.g. mate stamps are
+    /// dropped, not clamped; without that filter they are clamped into range).
+    /// Useful to normalise teacher files whose encode variants clip at
+    /// different ceilings. Must be in `[1, 32767]`.
+    #[arg(long, global = true, value_parser = clap::value_parser!(i16).range(1..))]
+    pub(crate) score_clamp_abs: Option<i16>,
 
     /// Inject weights from a quantised NNUE binary before training starts
     /// (pretrained start). The optimizer state (Ranger m/v/slow/step) is
